refactor(shared): added all resources exported on index
