+++ src/modules/budget-box/domain/interfaces/budget-box.repository.interface.ts
@@ -1,4 +1,4 @@
-import IBudgetBox from "@modules/shared/interfaces/budget-box-model.interface";
+import { IBudgetBox } from "@shared/index";
 import { IBaseRepository } from "types-ddd";
 import BudgetBoxAggregate from "../budget-box.aggregate";
