+++ src/modules/shared/index.ts
@@ -1,1 +1,6 @@
-export * from './common/error-messages/messages';
+export * from './utils/error-messages/messages';
+export * from './interfaces/budget-box-model.interface';
+export * from './interfaces/entity-mock.interface';
+export * from './interfaces/reason-model.interface';
+export * from './interfaces/transaction-model.interface';
+export * from './interfaces/user-model.interface';
