+++ src/modules/user/domain/interfaces/user.repository.interface.ts
@@ -1,5 +1,5 @@
 import { IBaseRepository } from 'types-ddd';
 import { UserAggregate } from '../user.aggregate';
-import IUser from '@shared/interfaces/user-model.interface';
+import { IUser } from '@shared/index';
 
 export type IUserRepository = IBaseRepository<UserAggregate, IUser>;
